Split WorkerGoals and WorkerMonitor
They are now too complex and should be put in dedicatd files.
