+++ src/scheduler/worker.rs
@@ -12,7 +12,7 @@ use crossbeam::queue::ArrayQueue;
 #[cfg(feature = "count_live_bytes_in_gc")]
 use std::sync::atomic::AtomicUsize;
 use std::sync::atomic::Ordering;
-use std::sync::{Arc, Condvar, Mutex, MutexGuard};
+use std::sync::{Arc, Mutex};
 
 /// Represents the ID of a GC worker thread.
 pub type ThreadId = usize;
@@ -33,8 +33,9 @@ pub fn current_worker_ordinal() -> ThreadId {
     ordinal
 }
 
-/// The part shared between a GCWorker and the scheduler.
-/// This structure is used for communication, e.g. adding new work packets.
+/// The struct has one instance per worker, but is shared between workers via the scheduler
+/// instance.  This structure is used for communication between workers, e.g. adding designated
+/// work packets, stealing work packets from other workers, and collecting per-worker statistics.
 pub struct GCWorkerShared<VM: VMBinding> {
     /// Worker-local statistics data.
     stat: AtomicRefCell<WorkerLocalStat<VM>>,
@@ -68,246 +69,6 @@ impl<VM: VMBinding> GCWorkerShared<VM> {
     #[cfg(feature = "count_live_bytes_in_gc")]
     pub(crate) fn get_and_clear_live_bytes(&self) -> usize {
         self.live_bytes.swap(0, Ordering::SeqCst)
-    }
-}
-
-/// The result type of the `on_last_parked` call-back in `WorkMonitor::park_and_wait`.
-/// It decides how many workers should wake up after `on_last_parked`.
-pub(crate) enum LastParkedResult {
-    /// The last parked worker should wait, too, until more work packets are added.
-    ParkSelf,
-    /// The last parked worker should unpark and find work packet to do.
-    WakeSelf,
-    /// Wake up all parked GC workers.
-    WakeAll,
-}
-
-/// A data structure for synchronizing workers.
-///
-/// -   It allows workers to park and unpark.  It keeps track of the number of workers parked, and
-///     allows the last parked worker to perform operations that can only be performed when all
-///     workers have parked.
-/// -   It allows mutators to notify workers to schedule a GC.
-pub(crate) struct WorkerMonitor {
-    /// The synchronized part.
-    sync: Mutex<WorkerMonitorSync>,
-    /// Notified if workers have anything to do.  That include any work packets available, and any
-    /// field in `sync.goals.requests` set to true.
-    have_anything_to_do: Condvar,
-}
-
-/// The synchronized part of `WorkerMonitor`.
-pub(crate) struct WorkerMonitorSync {
-    /// The total number of workers.
-    worker_count: usize,
-    /// Number of parked workers.
-    parked_workers: usize,
-    /// Current and requested goals.
-    goals: WorkerGoals,
-}
-
-#[derive(Default)]
-pub(crate) struct WorkerGoals {
-    /// What are the workers doing now?
-    pub(crate) current: Option<WorkerGoal>,
-    /// Requests received from mutators.
-    pub(crate) requests: WorkerRequests,
-}
-
-/// The thing workers are currently doing.  This affects several things, such as what the last
-/// parked worker will do, and whether workers will stop themselves.
-pub(crate) enum WorkerGoal {
-    Gc {
-        start_time: std::time::Instant,
-    },
-    #[allow(unused)] // TODO: Implement forking support later.
-    StopForFork,
-}
-
-/// Reqeusts received from mutators.  Workers respond to those requests when they do not have a
-/// current goal.  Multiple things can be requested at the same time, and workers respond to the
-/// thing with the highest priority.
-///
-/// The fields of this structs are ordered with decreasing priority.
-#[derive(Default)] // All fields should be false by default.
-pub(crate) struct WorkerRequests {
-    /// The VM needs to fork.  Workers should save their contexts and exit.
-    pub(crate) stop_for_fork: bool,
-    /// GC is requested.  Workers should schedule a GC.
-    pub(crate) gc: bool,
-}
-
-impl WorkerMonitor {
-    pub fn new(worker_count: usize) -> Self {
-        Self {
-            sync: Mutex::new(WorkerMonitorSync {
-                worker_count,
-                parked_workers: 0,
-                goals: Default::default(),
-            }),
-            have_anything_to_do: Default::default(),
-        }
-    }
-
-    /// Request a GC worker to schedule the next GC.
-    /// Callable from mutator threads.
-    pub fn request_schedule_collection(&self) {
-        let mut guard = self.sync.lock().unwrap();
-        if !guard.goals.requests.gc {
-            guard.goals.requests.gc = true;
-            self.notify_work_available_inner(false, &mut guard);
-        }
-    }
-
-    /// Wake up workers when more work packets are made available for workers,
-    /// or a mutator has requested the GC workers to schedule a GC.
-    pub fn notify_work_available(&self, all: bool) {
-        let mut guard = self.sync.lock().unwrap();
-        self.notify_work_available_inner(all, &mut guard);
-    }
-
-    /// Like `notify_work_available` but the current thread must have already acquired the
-    /// mutex of `WorkerMonitorSync`.
-    fn notify_work_available_inner(&self, all: bool, _guard: &mut MutexGuard<WorkerMonitorSync>) {
-        if all {
-            self.have_anything_to_do.notify_all();
-        } else {
-            self.have_anything_to_do.notify_one();
-        }
-    }
-
-    /// Park a worker and wait on the CondVar `work_available`.
-    ///
-    /// If it is the last worker parked, `on_last_parked` will be called.
-    /// The argument of `on_last_parked` is true if `sync.gc_requested` is `true`.
-    /// The return value of `on_last_parked` will determine whether this worker and other workers
-    /// will wake up or block waiting.
-    pub fn park_and_wait<VM, F>(&self, worker: &GCWorker<VM>, on_last_parked: F)
-    where
-        VM: VMBinding,
-        F: FnOnce(&mut WorkerGoals) -> LastParkedResult,
-    {
-        let mut sync = self.sync.lock().unwrap();
-
-        // Park this worker
-        let all_parked = sync.inc_parked_workers();
-        trace!(
-            "Worker {} parked.  parked/total: {}/{}.  All parked: {}",
-            worker.ordinal,
-            sync.parked_workers,
-            sync.worker_count,
-            all_parked
-        );
-
-        let mut should_wait = false;
-
-        if all_parked {
-            debug!("Worker {} is the last worker parked.", worker.ordinal);
-            let result = on_last_parked(&mut sync.goals);
-            match result {
-                LastParkedResult::ParkSelf => {
-                    should_wait = true;
-                }
-                LastParkedResult::WakeSelf => {
-                    // Continue without waiting.
-                }
-                LastParkedResult::WakeAll => {
-                    self.notify_work_available_inner(true, &mut sync);
-                }
-            }
-        } else {
-            should_wait = true;
-        }
-
-        if should_wait {
-            // Notes on CondVar usage:
-            //
-            // Conditional variables are usually tested in a loop while holding a mutex
-            //
-            //      lock();
-            //      while condition() {
-            //          condvar.wait();
-            //      }
-            //      unlock();
-            //
-            // The actual condition for this `self.work_available.wait(sync)` is:
-            //
-            // 1.  any work packet is available, or
-            // 2.  a request for scheduling GC is submitted.
-            //
-            // But it is not used like the typical use pattern shown above, mainly because work
-            // packets can be added without holding the mutex `self.sync`.  This means one worker
-            // can add a new work packet (no mutex needed) right after another worker finds no work
-            // packets are available and then park.  In other words, condition (1) can suddenly
-            // become true after a worker sees it is false but before the worker blocks waiting on
-            // the CondVar.  If this happens, the last parked worker will block forever and never
-            // get notified.  This may happen if mutators or the previously existing "coordinator
-            // thread" can add work packets.
-            //
-            // However, after the "coordinator thread" was removed, only GC worker threads can add
-            // work packets during GC.  Parked workers (except the last parked worker) cannot make
-            // more work packets availble (by adding new packets or opening buckets).  For this
-            // reason, the **last** parked worker can be sure that after it finds no packets
-            // available, no other workers can add another work packet (because they all parked).
-            // So the **last** parked worker can open more buckets or declare GC finished.
-            //
-            // Condition (2), i.e. `sync.should_schedule_gc` is guarded by the mutator `sync`.
-            // When set (by a mutator via `request_schedule_collection`), it will notify a
-            // worker; and the last parked worker always checks it before waiting.  So this
-            // condition will not be set without any worker noticing.
-            //
-            // Note that generational barriers may add `ProcessModBuf` work packets when not in GC.
-            // This is benign because those work packets are not executed immediately, and are
-            // guaranteed to be executed in the next GC.
-
-            // Notes on spurious wake-up:
-            //
-            // 1.  The condition variable `work_available` is guarded by `self.sync`.  Because the
-            //     last parked worker is holding the mutex `self.sync` when executing
-            //     `on_last_parked`, no workers can unpark (even if they spuriously wake up) during
-            //     `on_last_parked` because they cannot re-acquire the mutex `self.sync`.
-            //
-            // 2.  Workers may spuriously wake up and unpark when `on_last_parked` is not being
-            //     executed (including the case when the last parked worker is waiting here, too).
-            //     If one or more GC workers spuriously wake up, they will check for work packets,
-            //     and park again if not available.  The last parked worker will ensure the two
-            //     conditions listed above are both false before blocking.  If either condition is
-            //     true, the last parked worker will take action.
-            sync = self.have_anything_to_do.wait(sync).unwrap();
-        }
-
-        // Unpark this worker.
-        sync.dec_parked_workers();
-        trace!(
-            "Worker {} unparked.  parked/total: {}/{}.",
-            worker.ordinal,
-            sync.parked_workers,
-            sync.worker_count,
-        );
-    }
-}
-
-impl WorkerMonitorSync {
-    /// Increase the packed-workers counter.
-    /// Called before a worker is parked.
-    ///
-    /// Return true if all the workers are parked.
-    fn inc_parked_workers(&mut self) -> bool {
-        let old = self.parked_workers;
-        debug_assert!(old < self.worker_count);
-        let new = old + 1;
-        self.parked_workers = new;
-        new == self.worker_count
-    }
-
-    /// Decrease the packed-workers counter.
-    /// Called after a worker is resumed from the parked state.
-    fn dec_parked_workers(&mut self) {
-        let old = self.parked_workers;
-        debug_assert!(old <= self.worker_count);
-        debug_assert!(old > 0);
-        let new = old - 1;
-        self.parked_workers = new;
     }
 }
